pyproject.toml
[tool.poetry]
name = "pricer"
version = "0.1.0"
description = ""
authors = ["ngmysamuel <e0148675@u.nus.edu>"]
readme = "README.md"
packages = [{include = "pricer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
alpaca-py = "^0.43.2"
pandas = "^2.3.3"
streamlit = "^1.52.2"
seaborn = "^0.13.2"
matplotlib = "^3.10.8"
scipy = "^1.16.3"
yfinance = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
